+++ decidim-core/lib/decidim/core/test/factories.rb
@@ -153,6 +153,10 @@ FactoryGirl.define do
       admin { true }
     end
 
+    trait :user_manager do
+      roles { ["user_manager"] }
+    end
+
     trait :process_admin do
       transient do
         participatory_process { create(:participatory_process) }
@@ -181,6 +185,28 @@ FactoryGirl.define do
                participatory_process: evaluator.participatory_process,
                role: :collaborator
       end
+    end
+
+    trait :process_moderator do
+      transient do
+        participatory_process { create(:participatory_process) }
+      end
+
+      organization { participatory_process.organization }
+
+      after(:create) do |user, evaluator|
+        create :participatory_process_user_role,
+               user: user,
+               participatory_process: evaluator.participatory_process,
+               role: :moderator
+      end
+    end
+
+    trait :managed do
+      email { "" }
+      password { "" }
+      password_confirmation { "" }
+      managed { true }
     end
   end
 
@@ -336,5 +362,11 @@ FactoryGirl.define do
     moderation
     user { build(:user, organization: moderation.reportable.organization) }
     reason "spam"
+  end
+
+  factory :impersonation_log, class: Decidim::ImpersonationLog do
+    admin { build(:user, :admin) }
+    user { build(:user, :managed, organization: admin.organization) }
+    started_at Time.current
   end
 end